requirements.txt

pandas==1.3.3
pyspark==3.1.2
kafka-python==2.0.2
pymongo==3.12.0
scikit-learn==0.24.2
